refactor: 파일 분리, 예외처리 (#252)
