+++ frontend/src/components/ProfileTest/UserContainer/UserInfoEditModal/index.jsx
@@ -13,13 +13,36 @@ import { useHistory } from "react-router";
 export default function UserInfoEditModal({ setModalSwitch }) {
   const history = useHistory();
   const [userData, setUserData] = useState({});
+  const [newNickname, setNewNickname] = useState();
+  const [user, setUser] = useRecoilState(userState);
+  const [schoolId, setSchoolId] = useState(user.schoolId);
 
   const getUserData = async () => {
     await getMyInfo().then((res) => {
       setUserData(res.data.data);
     });
   };
-
+  const onChange = (e) => {
+    const { value, name } = e.target;
+    const copy = { ...userData };
+    copy[name] = value;
+    setUserData(copy);
+  };
+  const onFileChange = (e) => {
+    const copy = { ...userData };
+    switch (e.target.id) {
+      case "reset":
+        copy.avatar = null;
+        setUserData(copy);
+        break;
+      case "upload":
+        copy.avatar = e.target.files[0];
+        setUserData(copy);
+        break;
+      default:
+        return;
+    }
+  };
   const checkingNickname = (nickname) => {
     nickCheckApi(nickname)
       .then((res) => alert("사용가능한 아이디 입니다"))
@@ -29,12 +52,16 @@ export default function UserInfoEditModal({ setModalSwitch }) {
   const submitEditUser = (userData, schoolId, newNickname) => {
     editMyInfo(userData, schoolId, newNickname)
       .then((res) => {
+        const copy = { ...user };
+        copy.name = res.data.data.nickname;
+        copy.schoolId = res.data.data.schoolId;
+        setUser(copy);
         alert("수정이 완료되었습니다!");
         closeModal();
         history.push(`/profile/${res.data.data.nickname}`);
         history.go(0);
       })
-      .catch((err) => console.log(err));
+      .catch((err) => alert("수정 내용을 다시 한번 확인해주세요"));
   };
 
   const closeModal = () => {
@@ -60,9 +87,14 @@ export default function UserInfoEditModal({ setModalSwitch }) {
               <S.ModalContent>
                 <Presenter
                   userData={userData}
-                  setUserData={setUserData}
-                  submitEditUser={submitEditUser}
+                  onChange={onChange}
+                  onFileChange={onFileChange}
+                  newNickname={newNickname}
+                  setNewNickname={setNewNickname}
+                  schoolId={schoolId}
+                  setSchoolId={setSchoolId}
                   checkingNickname={checkingNickname}
+                  submitEditUser={submitEditUser}
                 />
               </S.ModalContent>
             </S.ModalBox>
